Generic Typing for SelectList bound value
Allows the binding of the intended output value type directly instead of capturing in a generic object to be projected later. This makes it easier to bind directly to an object field or property.

This is part of a custom version of the selectlist I am using in production. I would like to provide the rest of it if this is well received.